Organization, dronePos, data parser, and unsafe class mutation
I think that's what the last one's called. Anyway, it's totally unsafe, but it should work in context. Not that I can really test it. Because of that, I've had to merge Shape.py and Circle.py

The data parser is just a framework, and it does not have the proper equation for this test data (the tag was moved randomly). But it does find the deviation.

I really hope GitHub understands this commit...